AudioManager, Connections Handling
